Update Common-Issues.md
Removed WorldEdit measure for Stage 3 of Floating tags: if the other two stages didn't work WE won't be useful at all. Updated it with the possibility a ghost Regulator/Cargo Manager is there instead, respawning the hologram over and over again.
Moved the note from Unplaceable blocks to Floating tags: it's way more likely to confuse them this way for the reason above. Having it under Unplaceable blocks too would be overkill as it should be very easy to distinguish an invisible armor stand blocking placement from placement being cancelled.
Added a better way to deal with isolated ghost blocks: in this case debug fish only is faster (and allows for the user to also confirm the issue if they wish to as most of the times it happens with head items rather than blocks, at least from my personal experience).